requirements.txt
flask
requests
beautifulsoup4
pydub
python-telegram-bot>=20.0
yt-dlp
